requirements.txt
certifi==2021.5.30
numpy==1.21.1
pandas==1.3.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
wincertstore==0.2

matplotlib~=3.3.4
seaborn~=0.11.1
colorama~=0.4.4